refactor: centralize debounce implementation
